- add test for https://github.com/linq2db/linq2db/issues/4657
- move custom context tests to separate fixture
