+++ Tests/EntityFrameworkCore/ContextTestBase.cs
@@ -1,20 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Data.Common;
 using System.Linq;
-
-using LinqToDB.Data;
 
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Diagnostics;
 
 using Tests;
-
-#if NETFRAMEWORK
-using MySqlConnectionStringBuilder = MySql.Data.MySqlClient.MySqlConnectionStringBuilder;
-#else
-using MySqlConnectionStringBuilder = MySqlConnector.MySqlConnectionStringBuilder;
-#endif
 
 namespace LinqToDB.EntityFrameworkCore.Tests
 {
@@ -52,59 +43,6 @@ namespace LinqToDB.EntityFrameworkCore.Tests
 
 		protected virtual void OnDatabaseCreated(string provider, TContext context)
 		{
-		}
-
-		protected string GetConnectionString(string provider)
-		{
-			var efProvider = provider + ".EF";
-			var connectionString = DataConnection.TryGetConnectionString(efProvider);
-
-			if (connectionString == null)
-			{
-				var originalCS = connectionString = DataConnection.GetConnectionString(provider);
-				var dbProvider = DataConnection.GetDataProvider(provider);
-
-				// create and register ef-specific connection string
-				// and create database if needed (if EnsureCreated doesn't do it)
-				switch (provider)
-				{
-					case var _ when provider.IsAnyOf(TestProvName.AllSqlServer):
-					{
-						var cnb = new Microsoft.Data.SqlClient.SqlConnectionStringBuilder(connectionString);
-						cnb.InitialCatalog += ".ef";
-						connectionString = cnb.ConnectionString;
-						break;
-					}
-					case var _ when provider.IsAnyOf(TestProvName.AllPostgreSQL):
-					{
-						var cnb = new Npgsql.NpgsqlConnectionStringBuilder(connectionString);
-						cnb.Database += "_ef";
-						connectionString = cnb.ConnectionString;
-						break;
-					}
-					case var _ when provider.IsAnyOf(TestProvName.AllMySql):
-					{
-						var cnb = new MySqlConnectionStringBuilder(connectionString);
-						cnb.Database += "_ef";
-						cnb.PersistSecurityInfo = true;
-						connectionString = cnb.ConnectionString;
-						break;
-					}
-					case var _ when provider.IsAnyOf(TestProvName.AllSQLite):
-					{
-						var cnb = new Microsoft.Data.Sqlite.SqliteConnectionStringBuilder(connectionString);
-						cnb.DataSource = $"sqlite.{provider}.ef.db";
-						connectionString = cnb.ConnectionString;
-						break;
-					}
-					default:
-						throw new InvalidOperationException($"{nameof(GetConnectionString)} is not implemented for provider {provider}");
-				}
-
-				DataConnection.AddConfiguration(efProvider, connectionString, dbProvider);
-			}
-
-			return connectionString;
 		}
 
 		private void InitializeDatabase(TContext context, string provider, string connectionString)
